MTPSol/requirements.txt
fair-esm
biopython
torch_geometric
torch_scatter
biotite
matplotlib
pandas
scikit-learn
seaborn